+++ tests/src/lib.rs
@@ -544,8 +544,6 @@ fn run_tests(cfg: TestConfig) -> Result<bool, io::Error> {
                     sql::prepared_with_unnamed_params,
                     sql::prepared_with_named_params,
                     sql::prepared_invalid_params,
-                    tuple_picodata::tuple_format_get_names,
-                    tuple_picodata::tuple_as_named_buffer,
                     tuple_picodata::tuple_hash,
                 ])
             }